Renamed keysEquals.ts file
