+++ src/session/crud/putSession.ts
@@ -2,7 +2,7 @@ import { AppSession } from '../types'
 import { AuthHandlerParams } from '../../storyblok-auth-api'
 import { setAllSessions } from './setAllSessions'
 import { getAllSessions } from './getAllSessions'
-import { keysEquals } from './utils/keys-equals'
+import { keysEquals } from './utils/keysEquals'
 import { GetCookie, SetCookie } from '../../utils'
 
 export type PutSessionParams = Pick<
